Move token saving and loading to separate file
